requirements.txt
alembic==1.12.0
attrs==23.1.0
blinker==1.6.2
cattrs==23.1.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
flask-sqlalchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.1
platformdirs==3.10.0
psycopg2==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
requests-cache==1.1.0
six==1.16.0
SQLAlchemy==2.0.20
typing-extensions==4.7.1
url-normalize==1.4.3
urllib3==2.0.4
werkzeug==2.3.7
WTForms==3.0.1
zipp==3.16.2
